move NativeRequire.js to utils
