+++ app/Alspotify.ts
@@ -5,8 +5,8 @@ import alsong from 'alsong';
 import Koa from 'koa';
 import bodyParser from 'koa-bodyparser';
 import Router from 'koa-router';
-import nativeRequire from './NativeRequire';
 import utils, {ConfigApi} from './utils/Config';
+import nativeRequire from './utils/NativeRequire';
 import Observable, {Observer} from './utils/Observable';
 import fs from 'fs';
 import * as path from 'path';
