+++ app/utils/NativeRequire.js
